+++ lib/screens/passwords_overview_screen.dart
@@ -97,8 +97,11 @@ class _PasswordsOverviewScreenState
                       child: Scrollbar(
                         child: ListView.builder(
                           itemCount: passwords.length,
-                          itemBuilder: (ctx, i) =>
-                              PasswordListItem(passwords[i], deletePassword),
+                          itemBuilder: (ctx, i) => PasswordListItem(
+                            passwords[i],
+                            deletePassword,
+                            autofillMode,
+                          ),
                         ),
                       ),
                     ),